added TelegramUpdate save method to TelegramUpdateService and implemented it

+++ wrapper-ripple/src/main/java/com/hatiko/ripple/telegram/bot/core/repository/TelegramChatRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import com.hatiko.ripple.telegram.bot.core.model.TelegramChat;
 
 @RepositoryRestResource(collectionResourceRel = "telegram_chats", path = "chats")
-public interface TelegramChatRepository extends PagingAndSortingRepository<TelegramChat, Integer>{
+public interface TelegramChatRepository extends PagingAndSortingRepository<TelegramChat, Long>{
 	
 	Optional<TelegramChat> findByUser_Id(Integer userId);
